editor: don't desensitize dialog if connection isn't valid
Dialog sensitivity should be desensitized if a system connection is invalid,
otherwise the user wouldn't be able to fix the connection to make it valid
again.  For system connections, dialog should be insensitive only when the
user is (a) not authorized to edit the connection, or (b) the connection is
read-only.