gconf: fix connection update signal emission
After a change was made (in the connection editor or whatever) the applet
wasn't sending out an updated signal.  Furthermore, filling the default
VPN user name was causing unwanted GConf updates, so we now fill it only
when sending the connection over dbus, which is the only time it's needed
anyway.