editor: fix PolicyKit button refcounting
When creating GTK widgets manually, they start with a floating
reference not a real reference, so we need to convert that
floating ref into a real ref before really using it, or we'll get
duplicate calls to the dispose method.  glade usually handles
this behind-the-scenes for us, but when making our own widgets we
have to do that manually.