editor: fix editing of system VPN connections
All system connections fire off a Secrets request right before
their UI is shown; and in the case of VPN connections, the UI
wasn't constructed until that Secrets call returned.  But the
core editor code expected it to be constructed right after the
call was kicked off, and so the VPN tab wouldnt' be shown for any
system VPN connection.