vpn: add user name for connection update signals too
The username wasn't getting added for connection update signals, just
for GetSettings, so after a successful connection the applet would
update the timestamp, which triggers a connection update signal, which
wouldn't include the vpn username.  Since we have protections to
not write the username out to GConf, just ignore clearing the username
anyway after adding it too.