applet: assume system connections are valid in the wireless dialog
It's actually kinda pointless to request secrets from system connections
in the wireless dialog, because (a) the plaintext secrets are hidden from
the user by dots, and (b) the user can't change them here.  Since the
connection couldn't have been exported over D-Bus if it wasn't valid in the
first place, assume it is valid no matter what.