applet: assume system connections are valid in the wireless dialog
Since 0.7 doesn't request secrets from system connections, there wasn't
really any way to validate the connection, and the wireless dialog's
Connect... button would always be insensitive when picking a system
connection from the "Connect to hidden..." menu item.  Since the
connection couldn't have been exported over D-Bus if it wasn't
valid in the first place, assume it is valid even though we don't
have secrets for it.  This is fixed in NM 0.8.