tls: require password-protected private keys
We already required this, but unencrypted private keys got by the
cert chooser dialog's filter.  Fix that by alerting the user that
the key appears to be unencrypted and that they'll need to
encrypt it to use it.

Usually caused by outdated NM 0.6.x-era instructions for connecting
to eduroam, for example:

http://imt.uni-paderborn.de/unser-angebot/dienste-a-z/dienste-nach-themen/wlan/netzwerk-eduroam/netzwerkzertifikate/linux-gui/