editor: serialize PolicyKit secrets requests (rh #462944) (lp:462944)
PolicyKit doesn't queue up authorization requests internally.  Instead,
if there's a pending authorization request, subsequent requests for that
same authorization will return NotAuthorized+Challenge.  That's pretty
inconvenient and it would be a lot nicer if PK just queued up subsequent
authorization requests and executed them when the first one was finished.
But since it doesn't do that, we have to serialize the authorization
requests ourselves to get the right authorization result.

This can be reverted whenever (if) PK starts serializing the requests
internally.