applet: escape notification text
Notification messages are a subset of HTML, and thus have to have
characters like '&' escaped, otherwise the notification daemon
displays a blank alert.  Consolidate the escaping code so that
we can use it for both VPN banners and other notifications.  Ran
into this problem when the applet was trying to display a connect
notification on connection to the 'AT&T MEdia Net' connection
created by the mobile broadband wizard.