core: ensure user connection scope after 8627880e07c8345f69ed639325280c7f62a8f894
Previously the scope would only have been set when the connection was exported
but since the connection editor shouldn't ever export connections, the scope
has to be set manually.  Without this fix the editor will be confused as to
the scope of connections and will try to add a connection instead of updating
it.