core: fix clearing secrets when updating connections
It was never a problem for wifi/wired or DSL since their various
configurations require valid secrets before you can hit Apply.  But
for Mobile Broadband PINs the libnm-glib settings interface
refactor broke clearing secrets so you could never remove a PIN
from a 3G connection after you added one.

The problem was that empty keys were never deleted from the keyring
when updating the connection.  Make that happen.  But we have to be
careful because sometimes in the applet the connection won't have
secrets when it gets updated (timestamps, seen BSSIDs), so we don't
want to blow away keyring items we simply haven't asked for yet in
those two cases.