gsm: fix secret requests when connection has multiple secrets
We need	to return either *all* secrets in the secret request
reply (either only all the secrets, or the whole setting hash
including all the secrets).  But not a setting hash including
only *some* secrets like we were doing.  Otherwise if there are
multiple secrets (like a GSM password and a PIN) we'll get into
a loop asking for each secret in turn, but never actually using
them correctly.