core: fix "Always ask for this password" behavior
The bug is that for 802.1x connections using simple EAP methods
(like TTLS or PEAP) the applet would continually ask for the
password even though "Always ask" was not checked.

The applet still needs the password to pass back to NM when NM
needs it to make a connection, so we only want to desensitize the
password entry and clear out the password in nm-connection-editor
when "Always ask" is checked.

On the applet side, when the connection is an 'always ask' connection,
store the "always ask" state, but still accept password entry and
send it along to NM in the GetSecrets response when the user hits OK.