editor: fix issues with multiple concurrent operations (rh #585405)
The editor was allowing itself to be closed while it was awaiting
authorization for updates.  That's bad since (as yet) we don't
have a way to cancel an ongoing operation that's blocked for
authorization; that would require API changes to libnm-glib. So
instead, just don't allow interaction with the editor until the
user has completed or canceled authorization.

Also fixes a small bug where the connection's scope would not be
reset if the user canceled authorization.