applet: make more robust against wireless dialog secrets response (rh #582428)
If the user cancels the wireless dialog while it's waiting for
authorization for secrets or something, the user could have clicked
the cancel button and closed the dialog.  Then when the	secrets	request
comes back the dialog is closed and we go boom when we try to
access the dead dialog's private members.

So ensure that closing the dialog results in the secrets callback
not trying to access the dialog at all, and make the buttons
insensitive while the secrets call is going on.