gconf: remove fully deleted connections (rh #618973)
Sometimes the GConf directory will already be deleted by the time
the notify-consolidator	runs, and before this patch the	applet
simply wouldn't	pick up	that case because it checked for the
connection's GConf directory before treating the connection as
really deleted.  Even when the directory is completely gone we
want to make sure we forget about the connection entirely.