editor: make sure editor windows are destroyed (rh #572466)
The GladeXML object we get a top-level window or dialog from
doesn't actually keep the reference to the object, we own it
after the glade_xml_get_widget() call.  So we need to destroy
it, otherwise we'll keep getting signals from the CEPolkitButton
in the editor window after we've already destroyed it's controlling
NMConnectionEditor object, leading to badness.