editor: correction of CEPolkitButton handling during permissions call (rh #603566)
Commit 5978047977a17f35d9a3961023143506a73899f8 cause crashes of nm-connection-editor
when changing a system connection (clicking Apply button). It didn't add PermInfo
structure to perm_calls list.
Even if this commit fixes the crash, the code isn't still 100% correct because the
get_permissions_cb() gets called after finalize() (and thus info->disposed as well as
the whole object should not be used).