gsm/cdma: MCC/MNC and SID are provider properties, not method ones
And some providers can have roaming-only networks without native
service, thus would have an MCC/MNC which we'd like to use, but
won't have an APN.  The MCC/MNC and SID are really provider/network
properties, not plan/APN properties.

As a bonus, this patch saves memory because the MCC/MNC list doesn't
have to be copied for each plan.