applet: fix menu lifetime issues and GTK 2.91 drawing problems
On recent GTK, the menu would be shown briefly then disappear, because
the menu destroy callback was being triggered when the menu was displayed
due to signal timing differences in new GTK.  All that was stupid anyway
so clean up the menu lifetime and simplify display and destruction.

As a bonus, this fixes the menu getting leaked every so often.