applet: track keyring requests for possible cancelation
Since the keyring doesn't pass the request ID to the callback of
the request, we have to track it via a small allocated structure.
Yay for the keyring.

Previously the SaveSecrets handler would overwrite the keyring id
of a previous operation if multiple secrets were to be saved.  This
fixes that issue.