wifi: fix wireless dialog for secret agent operation
When the wifi dialog is called to get secrets when NM requests
them, the connection given won't be an NMRemoteConnection, and
it'll already be populated with all available secrets.  So we
don't need to get more.

Second, be permissive about the device when getting the updated
connection when the dialog is finished.  Some callers might not
care.