vpn: send whole connections to VPN auth dialogs
Since VPN connections are no longer stored just in GConf, we need
to send all the VPN details to the plugins' auth dialog.  The
protocol between the auth-dialog and the applet needs to be updated
for this, so VPN plugins will need modification, but in return
they can rip out a lot of code that talks to GConf.

Furthermore, send secrets back to NM as a GHashTable of string:string
like the NMSettingVPN 'secrets' property acutally is; it can be done
more simply than before.