applet: pass '-i' flag to VPN auth dialogs to signal interaction is allowed
Sometimes we just want the auth dialog to return any existing secrets
it knows about, but not prompt the user for anything, which is what
should happen if this flag is not passed.  Only if the flag is passed
should the VPN prompt the user for any input.

To clarify "interaction allowed vs. retry", retry means that the old
secrets aren't good and completely new ones should be requested, but
retry still requires "interaction allowed" before the user should
be prompted for anything.  The applet should never pass "retry"
without "interaction allowed".