applet: more import fixes for vpnc and pptp
A historical oddity with vpnc is that the passwords are stored in
the keyring using different names than the NM-vpnc property names
for them; "password" == "Xauth password", while "group-password" ==
"IPSec secret".

So after the manual VPN-specific setting of secret flags, don't
run the generic code below, because the generic code would find
the "group-password" or "password" items in the keyring, and add a
secret flag value for them, which would not be a known NM-vpnc
property and thus would fail validation in the plugin daemon.