editor: fix widget refcounting by cleanup up .ui toplevel widgets
Since GtkBuilder constructs all the widgets in the .ui file when
using gtk_builder_load_from_file(), every toplevel gets constructed
and then added to the 'toplevel_list' internally in GTK.  If you
don't then explicitly destroy every toplevel that got constructed
by the builder, the toplevel you forgot about still exists and
is still part of the "toplevel_list" leading to possible crashes
later on GTK issues grabs for eg list scrolling.

The editor and applet made heavy use of Glade's "root" argument when
loading .glade files to only construct a part of the widget
hierarchy, but those child widgets were always embedded in a toplevel
due to previous limitations of Glade 2.x.  But when converting over
to GtkBuilder, the code still kept ignoring the toplevel the child
was embedded in, even though we now need to track and destroy that
toplevel.  Thus, when the code was done with its child widget and
destroyed that child widget, the ignored toplevel still thought it
had children, and when GTK issued grabs due to scrolling the
live-but-hidden toplevel tried to execute operations on its
long-dead children.

We can of course use gtk_builder_add_objects_from_file() to emulate
the 'root' argument of Glade, but that's somewhat cumbersome, and
performance is better if we just split out toplevel widgets into
their own .ui files anyway, and get rid of toplevels from .ui files
where they aren't needed.