wifi: fix connections to new 802.1x/Enterprise networks (bgo #646210)
After the refactor for the Agent stuff, 802.1x got broken because the
get_more_info hook wasn't called.  Second, we don't need to build up
the whole connection anymore since NM can do that for us.  So fix stuff
such that when connecting to 802.1x networks we do show the dialog to
get the information we need, then we ship that off to NM to get
completed and connected.