applet: set password flags for old connections too
If the connection hasn't been edited in a long time (since before
password "types" came about in vpn) then we wouldn't have set the
password flags accordingly, because the code doesn't actually
check the keyring for the password at all (as that might trigger
a keyring unlock just for running the applet which sucks).  Make
sure we set the right password flags even if there isn't an old
"type" field in the data.