agent: save updated secrets before sending to NetworkManager
When updated secrets are retrieved during a GetSecrets we need to
save them, since we are the agent, and the agent is expected to
save secrets it controls.  NetworkManager will not send a SaveSecrets
request for agent-owned secrets as a result of a GetSecrets request.

This ensures that agent-owned VPN and 802.1x secrets (and any other
agent-owned secrets) are saved to GConf when entered by the user such
as when the secret is requested for the first time.