editor: fix compat with GTK+ <= 2.22
Which doesn't have GtkComboBoxText, so we have to use the old way.
It gets a bit complicated because we can't use the GtkBuilder file
to store the widget (since that can't change programmatically) so
we replace the widget in the Builder file with an alignment that
we stick the programmatically created combo box into.