po: return po/iso3166.tab.h back into POTFILES.in
po/iso3166.tab.h is a generated file containing country names to be translated.
The file is generated from /usr/share/zoneinfo/iso3166.tab by Rules-iso3166,
that gets appended to Makefile.
Rules-iso3166 is now updated to search for the input file in
$(datadir)/zoneinfo/iso3166.tab, and if not found in /usr/share/zoneinfo/iso3166.tab
When no input file is found, empty po/iso3166.tab.h is created and a warning
printed.