editor: fix interaction of single-edit mode with PolicyKit authentication (bgo #667356)
The editor wasn't properly waiting for PolicyKit authentication to
complete before quitting if it was launched from the GNOME Control
Center network applet with --uuid xxxxx.  So only emit the "editing-done"
signal when we're actually done with the update.

Second this fixes a possible double-free in the update error case
where connection_updated_cb() would free 'info' and so would
the edit_done_cb() right after calling connection_updated_cb().