editor: fix defaults for lcp-echo-failure and lcp-echo-interval (bgo #663970)
The default values have to be set in dsl_connection_new() so that the change
only applies for new connections. ce_page_ppp_new() is called every time a
connection is edited and thus we can't change the properties values in it. Else
"Send PPP echo packets" won't be initialized properly for connections that have
PPP settings all default and thus no PPP setting was present.