connection-editor: ensure that pages always appear in the correct order
If the connection-editor had to ask for authentication in order to
retrieve the secrets for a given editor page, then that page would end
up getting appending to the notebook out of order. Fix it so that
pages always end up in the order we called add_page() in, regardless
of what order they call page_initialized() in.