libnm-gtk: import mobile provider related changes from gnome-shell
The code in gnome-shell was modified to better handle several introspection
related issues (e.g. avoid lists of objects returned as values of hash table
elements).

This commit aligns the requirements of gnome-shell directly in libnm-gtk.

Bits and pieces were fixed in the mobile broadband connection creator wizard and
in the CDMA/GSM items of the standard nm-applet.