libnm-gtk: make all mobile provider types opaque
All types defined in the mobile provider parsing code are now opaque, in case we
need to update them in the future.

'NMAGsmMccMnc' is the only one which is left with the members public, as it is
not really likely to be modified in the current format.

Also, stick to use G_DEFINE_BOXED_TYPE() to define the types, and make all
reference count operations thread-safe.