applet: simplify launching the PIN/PUK code request dialog
The generation of the PIN/PUK request dialog was split into two separate places:
  a) The generic creation of the dialog
  b) The setup of the text and contents to show in the dialog

Given that this dialog is generated only for the single purpose of requesting
PIN or PUK, just merge both code chunks.

This fix will simplify the integration of the new 'broadband' object for modems
exported by the new ModemManager interface.