bluetooth: more gracefully handle new and old ModemManager
Calling dun_error() is a terminal condition and the operation
should not proceed after that.  So only call dun_error() when
setting up our D-Bus service proxies if both new and old MM
can't be contacted.

Previously when the plugin was built with MM1 support, but MM1
wasn't running, an error message would be displayed but the
DUN setup would continue to run.