editor: don't overwrite master name with master UUID
This code was left over from when master interface names weren't
as easy to deal with in the client code.  Plus, it tries to set
the master property of the NMConnection, which doesn't have a
master property.  It really wanted to set the master property
of NMSettingConnection, but since we don't want to do that,
and it was broken anyway, just remove it.