applet: MM1 doesn't need to be running for broadband class methods (bgo #694676)
The applet device classes are used for a number of things from UI
display to device activation, not all of which require ModemManager
to be running.  In this case, the nature of IPC means that the
applet may notice that MM1 has quit before it gets the NM signal
that the mobile broadband device has been removed.  Then the NM
signal comes in, and the applet needs to ask the MB device class
what to do with that event, and we crash because the device class
no longer exists because MM1 isn't running.  But there's no good
reason to tie the device class to the lifetime of MM1.