applet: always show the applet
This is a patch that has been carried in Ubuntu for a while now. Instead of
hiding the applet if NetworkManager isn't running, make use of the messages
already included that mention that it's not, and just always keep showing
the applet; just with a different menu if NM is unavailable.