editor: hide BSSID for Ad-Hoc connections (rh #906133)
BSSID is also cleared when saving the Ad-Hoc connection so that it is not sent
to wpa_supplicant.

BSSID is random for Ad-Hoc networks. It is generated in kernel mac80211 subsystem
or driver (when Ad-Hoc network is created). So, there's no need to specify the
BSSID at all. Actually, wpa_supplicant fails to associate with bssid provided.
The code for generating BSSID for IBSS (Ad-Hoc) is here:
http://lxr.linux.no/linux+v3.7.6/net/mac80211/ibss.c#L685