applet: only request secrets via UI when allowed to
The user should only be presented with UI to enter secrets when
NetworkManager sets the ALLOW_INTERACTION flag in the secrets
request to the agent.  Otherwise, a connection editor requesting
existing secrets for display in its UI, for a connection which
doesn't have any secrets yet, could trigger the applet's secrets
dialog.