editor: remember/restore old password when toggle ask_password in wifi
In the wifi security dialog, when you set the "always ask" checkbox, the
password will be cleared. Remember the previous password and restore it when
you uncheck the box again.

This has the effect, that you can toggle this checkbox and don't have to
reenter the password.

This only works, as long as you don't change the "Security",
"Authentication" or "Inner authentication" type. This is
intentionally not to expose the password by accident.

Signed-off-by: Thomas Haller <thaller@redhat.com>