security: don't save CA cert ignore if editing is canceled
Previously the code would update GSettings whenever a security
input field changed and the provided security information had
to be re-validated. This meant the value would be saved even
if the user canceled editing of the connection.  Fix this.

Unfortunately, since only the upper-level code knows when the
editing has been canceled, the upper-level code has to handle
reading the initial values, and saving the new ones.  But that's
easy to do.

Signed-off-by: Thomas Haller <thaller@redhat.com>