connection-editor: fix display of vlan parent
For a VLAN with "parent" set to an ifname, the editor was mistakenly
showing the parent as being whatever entry for that device came
alphabetically first in the popup menu (which I had never noticed
before because I didn't have any ethernet connection names that came
alphabetically before the MAC address of my ethernet card).

Also, fix the parent menu to always list ifname-based parents before
connection-uuid-based parents, and change the description of the
connection-based ones a bit.