editor: initialize the destroy handler at last in eap_method_init/wireless_security_init
When 'eap_method_init' or 'wireless_security_init' fails to read the
UI xml file, it calls 'eap_method_unref' or 'wireless_security_unref',
respectively. The unref methods call the destroy handler, but they
should not do it in this case, because the objects were not fully
constructed.

Fix this, be setting the destroy handler only at the end of
the init methods.

Signed-off-by: Thomas Haller <thaller@redhat.com>