editor: change JSON Team/TeamPort config widget from GEntry to GTextView
and allow importing the config from a file.

libteam configuration is a string that can be quite long, and simple GEntry is
not much appropriate for it. This commit uses GTextView instead, so that
structured multiline JSON config can be typed. In addition to that, the import
button allows selecting a file whose content will be imported as the config
text.