applet: initialize ca_cert_ignore checkbox for "Connect to Hidden"
When opening the "Connect to Hidden Network" dialog for a connection
without CA authority, the ca_cert_ignore checkbox was not properly loaded
and always stayed unchecked.

With this change, EAP connections without ca_cert_ignore flag will be
validate as erroneous and the "Connect" button in the "Connect to Hidden
Network" dialog stays disabled.

Previously, the behaviour was indeterministic, and also it might trigger
an assert in eap_method_ca_cert_not_required_toggled.

Signed-off-by: Thomas Haller <thaller@redhat.com>