applet: remove wrong assert in eap_method_ca_cert_not_required_toggled
With the previous commit, the "Connect" button in the "Connect to Hidden
Network" is disabled, if the EAP connection does not validate.
Before, the user was able to click on the "Connect" button and the
assert in eap_method_ca_cert_not_required_toggled failed.

Still, the assert is too strict, so remove it.

Also fix an invalid warning in eap-method-tls.c

Signed-off-by: Thomas Haller <thaller@redhat.com>