applet: redo the gnome-shell interaction code
Change the shell interaction code so that we only create an
NMSecretAgent if the shell is not running, and we permanently destroy
it if the shell ever starts up.

Revert the change to not run nm-applet under GNOME 3 now, since it
should be safe to do so, and this solves the problem for users who
need the GSM PIN unlock dialog.

https://bugzilla.gnome.org/show_bug.cgi?id=709419