editor: generalize ce_spin_output_with_default()
We may want to use this for default value strings other than
"automatic" in the future, so change the name to "automatic" and
pass the string to the generic helper that does the actual work.

Also, the return value for the "output" signal for the spinbutton
should be 'gboolean' not 'gint'.