editor: populate connection right before saving it
Previously the only way the UI pages could update the connection's
settings with the new values from the UI was whenever any widget
changed, which required connecting all the widgets to the the
ce_page_changed() handler to re-run the validate method, which
populated the connection and validated it.

Not all widgets need to be validated, especially if the have
GtkAdjustments or are GtkToggleButtons, since these already enforce
valid values.

Re-validating the connection right before saving it allows these
widgets to not be hooked up to ce_page_changed() every time they
change, but still allows them to push their latest value to the
connection/setting right before the changes are saved.