editor: store secrets by default only for current user (rh #879566)
Storing secrets system-wide can be a security breach. Because such a secret is
available to all users on the system.

So we default to NM_SETTING_SECRET_FLAG_AGENT_OWNED flag for all passwords,
regardless whether the connection profile itself is available to all users or
not. And we add a popup menu for password entries that allows users to override
password storage. This is useful e.g. for the case when a user (admin) needs
the connection profile to be usable at boot time.

There and two pieces of data to store:
profile storage/availability - it is influenced with "Available to all users"
secret  storage/availability - it is influenced with secret flags
This commit decouples these two things, so that they are independent now.

Test case:
Alice creates a connection profile for her WPA Wi-Fi access point. When she
logouts, Bobs comes in, logs into his own account and he is able to connect to
Alice's AP, because the profile contains the password, and it is available for
Bob (and all users).

https://bugzilla.redhat.com/show_bug.cgi?id=879566