applet: fix crash on error with current glib
When NMApplet encountered an error at construct time, it would unref
itself and return NULL. glib 2.39 will g_error() out in this case,
because you're not supposed to do that, and because it used to cause
mysterious crashes later on if you did. (nm-applet never hit those
crashes because if nm_applet_new() returns NULL it exits, so it didn't
keep running long enough to hit the crash.)

Anyway, fix this by using GInitable to initialize NMApplet rather than
trying to do it from constructor().

https://bugzilla.redhat.com/show_bug.cgi?id=1034500